Mock test database

+++ internal/storages/user/model/user.go
@@ -4,5 +4,5 @@ package model
 type User struct {
 	ID       string `json:"user_id"`
 	Password string `json:"password"`
-	MaxTodo  int64  `json:"max_todo"`
+	MaxTodo  int  `json:"max_todo"`
 }
